Refactor(DK-323): mutate hook 전용 폴더 생성 및 이동
